* refactor javascript * new dao for tag colors * always recreate and not minify js and css in debug mode * dao backend usage with magic function

+++ www/helpers/View.php
@@ -93,24 +93,53 @@ class View {
      * @return void
      */
     public function genMinifiedJsAndCss() {
-        // minify js
+        if(\F3::get('DEBUG')!=0) {
+			
+		}
+		
+		// minify js
         $targetJs = \F3::get('BASEDIR').'/public/all.js';
-        //if(!file_exists($targetJs)) {
+        if(!file_exists($targetJs) || \F3::get('DEBUG')!=0) {
             $js = "";
             foreach(\F3::get('js') as $file)
-                $js = $js . "\n" . \JSMin::minify(file_get_contents(\F3::get('BASEDIR').'/'.$file));
+                $js = $js . "\n" . $this->minifyJs(file_get_contents(\F3::get('BASEDIR').'/'.$file));
             file_put_contents($targetJs, $js);
-        //}
+        }
     
         // minify css
         $targetCss = \F3::get('BASEDIR').'/public/all.css';
-        //if(!file_exists($targetCss)) {
+        if(!file_exists($targetCss) || \F3::get('DEBUG')!=0) {
             $css = "";
             foreach(\F3::get('css') as $file)
-                $css = $css . "\n" . \CssMin::minify(file_get_contents(\F3::get('BASEDIR').'/'.$file));
+                $css = $css . "\n" . $this->minifyCss(file_get_contents(\F3::get('BASEDIR').'/'.$file));
             file_put_contents($targetCss, $css);
-        //}
+        }
     }
     
+	
+	/**
+     * minifies javascript if DEBUG mode is disabled
+     *
+     * @return minified javascript
+	 * @param javascript to minify
+     */
+	private function minifyJs($content) {
+		if(\F3::get('DEBUG')!=0) 
+			return $content;
+		return \JSMin::minify($content);
+	}
+	
+	
+	/**
+     * minifies css if DEBUG mode is disabled
+     *
+     * @return minified css
+	 * @param css to minify
+     */
+	private function minifyCss($content) {
+		if(\F3::get('DEBUG')!=0) 
+			return $content;
+		return \CssMin::minify($content);
+	}
     
 }
